file transfers enhancements

+++ dms-core/src/com/ogya/dms/core/intf/listeners/DmsFileServer.java
@@ -4,6 +4,6 @@ import java.nio.file.Path;
 
 public interface DmsFileServer {
 
-	Path fileRequested(Integer fileId);
+	Path fileRequested(Long fileId);
 
 }
